refactor; checkIfItemInDatabase deleted, cartprovider renamed to UserProductStateContext
