+++ app/favorite/page.tsx
@@ -1,12 +1,12 @@
 "use client";
 import React, { useEffect, useState } from "react";
-import { createClient } from "@/utils/supabase/client";
-import { useCart } from "../context/CartContext";
+
 import { getFavoriteWithUserId, getProduct } from "../actions";
 import FavoriteContent from "@/components/FavoriteContent";
+import { useProductState } from "../context/UserProductStateProvider";
 
 const FavoritePage = () => {
-  const { favorite ,userId} = useCart(); // Get favorites from context
+  const { favorite ,userId} = useProductState(); // Get favorites from context
   const [products, setProducts] = useState<any>();
